refactor: split province overlay unit section parts (#3878)
Extract military, civilian/naval, and close-button library parts from the
province overlay unit-sections host so each part stays under file-size targets;
document the new parts in repo-and-packages.md.

Refs #3878
