--- app/lib/features/game/widgets/province_overlay/province_sea_zone_detail_overlay_unit_sections.dart
@@ -1,282 +1,0 @@
-
-part of 'province_sea_zone_detail_overlay.dart';
-
-Widget _buildMilitarySectionByOwner({
-  required AppLocalizations l10n,
-  required Game game,
-  required List<Unit> military,
-  required String humanPlayerId,
-  required String provinceId,
-  required Orders draftOrders,
-}) {
-  final pending = provincePanelPendingMilitaryLines(
-    game: game,
-    orders: draftOrders,
-    provinceId: provinceId,
-    humanPlayerId: humanPlayerId,
-    l10n: l10n,
-  );
-  if (military.isEmpty && pending.isEmpty) {
-    return _buildSection(l10n.provinceOverlay_sectionMilitary, _emptyBodyDashText());
-  }
-  if (military.isEmpty) {
-    return _buildSection(
-      l10n.provinceOverlay_sectionMilitary,
-      Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        mainAxisSize: MainAxisSize.min,
-        children: pending
-            .map(
-              (line) => Padding(
-                padding: const EdgeInsets.only(left: CtSpacing.m / 2),
-                child: Text(
-                  line,
-                  style: TextStyle(color: EditorialMonoclePalette.muted),
-                ),
-              ),
-            )
-            .toList(),
-      ),
-    );
-  }
-  final byOwner = <String, List<Unit>>{};
-  for (final u in military) {
-    byOwner.putIfAbsent(u.ownerId, () => []).add(u);
-  }
-  final ownerIds = byOwner.keys.toList()
-    ..sort((a, b) {
-      if (a == humanPlayerId) return -1;
-      if (b == humanPlayerId) return 1;
-      return _ownerName(l10n, game, a).compareTo(_ownerName(l10n, game, b));
-    });
-  return _buildSection(
-    l10n.provinceOverlay_sectionMilitary,
-    Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        ...ownerIds.map((oid) {
-          final list = byOwner[oid]!;
-          final byType = <String, int>{};
-          for (final u in list) {
-            byType[u.type] = (byType[u.type] ?? 0) + 1;
-          }
-          final name = _ownerName(l10n, game, oid);
-          return Padding(
-            padding: const EdgeInsets.only(bottom: CtSpacing.m),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  name,
-                  style: TextStyle(
-                    color: EditorialMonoclePalette.fg,
-                    fontWeight: FontWeight.w600,
-                  ),
-                ),
-                ...byType.entries.map((e) {
-                  final label = regimentTypeDisplayLabel(l10n, e.key);
-                  return Text(
-                    l10n.provinceOverlay_indentedCount(label, e.value),
-                    style: TextStyle(color: EditorialMonoclePalette.fg),
-                  );
-                }),
-              ],
-            ),
-          );
-        }),
-        if (pending.isNotEmpty) ...[
-          const SizedBox(height: CtSpacing.m / 2),
-          ...pending.map(
-            (line) => Padding(
-              padding: const EdgeInsets.only(left: CtSpacing.m / 2),
-              child: Text(
-                line,
-                style: TextStyle(color: EditorialMonoclePalette.muted),
-              ),
-            ),
-          ),
-        ],
-      ],
-    ),
-  );
-}
-
-Widget _buildCivilianSectionFiltered({
-  required AppLocalizations l10n,
-  required Game game,
-  required List<Unit> civilian,
-  required String humanPlayerId,
-  required PlayerView playerView,
-  required Orders draftOrders,
-}) {
-  final visible = civilian
-      .where(
-        (u) => foreignCivilianVisibleToPlayer(
-          unit: u,
-          viewerPlayerId: humanPlayerId,
-          view: playerView,
-        ),
-      )
-      .toList();
-  if (visible.isEmpty) {
-    return _buildSection(l10n.provinceOverlay_sectionCivilian, _emptyBodyDashText());
-  }
-  final workList = draftOrders.workOrdersByPlayerId[humanPlayerId] ?? const [];
-  return _buildSection(
-    l10n.provinceOverlay_sectionCivilian,
-    Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      mainAxisSize: MainAxisSize.min,
-      children: visible.map((u) {
-        if (u.ownerId == humanPlayerId) {
-          WorkOrder? pending;
-          for (final o in workList) {
-            if (o.unitId == u.id) {
-              pending = o;
-              break;
-            }
-          }
-          if (pending != null) {
-            final targetLabel = workOrderTargetDisplayLabel(
-              l10n,
-              pending.target,
-            );
-            return Text(
-              l10n.provinceOverlay_unitTarget(u.type, targetLabel),
-              style: TextStyle(color: EditorialMonoclePalette.fg),
-            );
-          }
-          return Text(
-            l10n.provinceOverlay_unitTarget(
-              u.type,
-              unitStatusDisplayLabel(l10n, u.status),
-            ),
-            style: TextStyle(color: EditorialMonoclePalette.fg),
-          );
-        }
-        final o = _ownerName(l10n, game, u.ownerId);
-        return Text(
-          l10n.provinceOverlay_foreignUnitStatus(
-            o,
-            u.type,
-            unitStatusDisplayLabel(l10n, u.status),
-          ),
-          style: TextStyle(color: EditorialMonoclePalette.muted),
-        );
-      }).toList(),
-    ),
-  );
-}
-
-Widget _buildNavalSection({
-  required AppLocalizations l10n,
-  required Game game,
-  required List<Fleet> fleets,
-  required String humanPlayerId,
-  required Orders draftOrders,
-  String? pendingNavalPortProvinceId,
-}) {
-  final pending = pendingNavalPortProvinceId == null
-      ? const <String>[]
-      : provincePanelPendingNavalLines(
-          game: game,
-          orders: draftOrders,
-          provinceId: pendingNavalPortProvinceId,
-          humanPlayerId: humanPlayerId,
-          l10n: l10n,
-        );
-  return _buildSection(
-    l10n.provinceOverlay_sectionNaval,
-    Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        if (fleets.isEmpty && pending.isEmpty) _emptyBodyDashText(),
-        if (fleets.isNotEmpty)
-          ...fleets.map((f) {
-            final ownerName = _ownerName(l10n, game, f.ownerId);
-            final byType = <String, int>{};
-            for (final s in f.ships) {
-              byType[s.typeId] = (byType[s.typeId] ?? 0) + 1;
-            }
-            final fleetLabel = f.id == homeFleetIdFor(f.ownerId)
-                ? l10n.naval_homeFleetLabel
-                : l10n.naval_fleetLabel(f.id);
-            final shipParts = byType.entries
-                .map((e) {
-                  final label = shipTypeDisplayLabel(l10n, e.key);
-                  return '$label×${e.value}';
-                })
-                .join(', ');
-            return Text(
-              l10n.provinceOverlay_fleetSummary(
-                ownerName,
-                fleetLabel,
-                shipParts,
-              ),
-              style: TextStyle(color: EditorialMonoclePalette.fg),
-            );
-          }),
-        if (pending.isNotEmpty) ...[
-          const SizedBox(height: CtSpacing.m / 2),
-          ...pending.map(
-            (line) => Padding(
-              padding: const EdgeInsets.only(left: CtSpacing.m / 2),
-              child: Text(
-                line,
-                style: TextStyle(color: EditorialMonoclePalette.muted),
-              ),
-            ),
-          ),
-        ],
-      ],
-    ),
-  );
-}
-
-/// Pixel-art close control (non-Material) keyed for tests as
-/// [kOverlayCloseKey]. Border colour resolves to `--accent-dim` and the
-/// `×` glyph paints in `--muted` per
-/// SPEC/ui/province-sea-zone-detail-overlay.md § Dark-theme chrome.
-class _OverlayCloseButton extends StatelessWidget {
-  const _OverlayCloseButton({this.onClose});
-
-  static const Key kOverlayCloseKey = Key('overlay_close');
-
-  /// Width of the brass-toned border around the glyph (matches catalog 1 px).
-  static const double _borderWidth = 1;
-
-  /// Font size of the `×` glyph (preserved from prior chrome so the close
-  /// control retains its visual weight relative to the header title).
-  static const double _glyphFontSize = 18;
-
-  final VoidCallback? onClose;
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      key: kOverlayCloseKey,
-      onTap: onClose,
-      child: Container(
-        padding: const EdgeInsets.symmetric(
-          horizontal: CtSpacing.m,
-          vertical: CtSpacing.m / 2,
-        ),
-        decoration: BoxDecoration(
-          border: Border.all(
-            color: EditorialMonoclePalette.accentDim,
-            width: _borderWidth,
-          ),
-        ),
-        child: Text(
-          '×',
-          style: TextStyle(
-            fontSize: _glyphFontSize,
-            color: EditorialMonoclePalette.muted,
-          ),
-        ),
-      ),
-    );
-  }
-}
